--- a/.changeset/rude-dryers-fry.md
+++ b/.changeset/rude-dryers-fry.md
@@ -1,0 +1,5 @@
+---
+"stylelint": minor
+---
+
+Added `function-url-quotes` autofix