--- a/.changeset/rude-dryers-fry.md
+++ b/.changeset/rude-dryers-fry.md
@@ -2,4 +2,4 @@
 "stylelint": minor
 ---
 
-Added `function-url-quotes` autofix
+Added: `function-url-quotes` autofix