--- a/lib/rules/function-url-quotes/__tests__/index.js
+++ b/lib/rules/function-url-quotes/__tests__/index.js
@@ -713,6 +713,74 @@ testRule({
 testRule({
 	ruleName,
 	config: ['always'],
+	fix: true,
+
+	reject: [
+		{
+			code: '@import url(foo.css);',
+			fixed: '@import url("foo.css");',
+			message: messages.expected('url'),
+		},
+		{
+			code: 'a { background: url(foo.css); }',
+			fixed: 'a { background: url("foo.css"); }',
+			message: messages.expected('url'),
+		},
+	],
+});
+
+testRule({
+	ruleName,
+	config: ['never'],
+	fix: true,
+
+	reject: [
+		{
+			code: '@import url("foo.css");',
+			fixed: '@import url(foo.css);',
+			message: messages.rejected('url'),
+		},
+		{
+			code: '@import url( "foo.css" );',
+			fixed: '@import url( foo.css );',
+			message: messages.rejected('url'),
+		},
+		{
+			code: "@import url('foo.css');",
+			fixed: '@import url(foo.css);',
+			message: messages.rejected('url'),
+		},
+		{
+			code: "@import url( 'foo.css' );",
+			fixed: '@import url( foo.css );',
+			message: messages.rejected('url'),
+		},
+		// {
+		// 	code: '@document url("http://www.w3.org/");',
+		// 	fixed: '@document url(http://www.w3.org/);',
+		// 	message: messages.rejected('url'),
+		// },
+		// {
+		// 	code: '@document url( "http://www.w3.org/" );',
+		// 	fixed: '@document url( http://www.w3.org/ );',
+		// 	message: messages.rejected('url'),
+		// },
+		{
+			code: 'a { background: url("foo.css"); }',
+			fixed: 'a { background: url(foo.css); }',
+			message: messages.rejected('url'),
+		},
+		{
+			code: "a { background: url('foo.css'); }",
+			fixed: 'a { background: url(foo.css); }',
+			message: messages.rejected('url'),
+		},
+	],
+});
+
+testRule({
+	ruleName,
+	config: ['always'],
 	customSyntax: 'postcss-html',
 
 	reject: [