--- a/lib/rules/function-url-quotes/__tests__/index.js
+++ b/lib/rules/function-url-quotes/__tests__/index.js
@@ -755,16 +755,22 @@ testRule({
 			fixed: '@import url( foo.css );',
 			message: messages.rejected('url'),
 		},
-		// {
-		// 	code: '@document url("http://www.w3.org/");',
-		// 	fixed: '@document url(http://www.w3.org/);',
-		// 	message: messages.rejected('url'),
-		// },
-		// {
-		// 	code: '@document url( "http://www.w3.org/" );',
-		// 	fixed: '@document url( http://www.w3.org/ );',
-		// 	message: messages.rejected('url'),
-		// },
+		{
+			code: '@document url("http://www.w3.org/");',
+			fixed: '@document url(http://www.w3.org/);',
+			message: messages.rejected('url'),
+		},
+		{
+			code: '@document url( "http://www.w3.org/" );',
+			fixed: '@document url( http://www.w3.org/ );',
+			message: messages.rejected('url'),
+		},
+		{
+			code: '@document   url( "http://www.w3.org/" );',
+			fixed: '@document   url( http://www.w3.org/ );',
+			message: messages.rejected('url'),
+			description: 'multiple spaces after name',
+		},
 		{
 			code: 'a { background: url("foo.css"); }',
 			fixed: 'a { background: url(foo.css); }',
@@ -774,6 +780,12 @@ testRule({
 			code: "a { background: url('foo.css'); }",
 			fixed: 'a { background: url(foo.css); }',
 			message: messages.rejected('url'),
+		},
+		{
+			code: "a { background :   url('foo.css'); }",
+			fixed: 'a { background :   url(foo.css); }',
+			message: messages.rejected('url'),
+			description: 'multiple spaces in between',
 		},
 	],
 });