--- a/lib/rules/function-url-quotes/__tests__/index.js
+++ b/lib/rules/function-url-quotes/__tests__/index.js
@@ -722,8 +722,55 @@ testRule({
 			message: messages.expected('url'),
 		},
 		{
+			code: '@import url( foo.css );',
+			fixed: '@import url( "foo.css" );',
+			message: messages.expected('url'),
+		},
+		{
+			code: '@document url(http://www.w3.org/);',
+			fixed: '@document url("http://www.w3.org/");',
+			message: messages.expected('url'),
+		},
+		{
+			code: '@document url( http://www.w3.org/ );',
+			fixed: '@document url( "http://www.w3.org/" );',
+			message: messages.expected('url'),
+		},
+		{
+			code: '@document   url( http://www.w3.org/ );',
+			fixed: '@document   url( "http://www.w3.org/" );',
+			message: messages.expected('url'),
+			description: 'multiple spaces after name',
+		},
+		{
+			code: '@document url-prefix(http://www.w3.org/Style);',
+			fixed: '@document url-prefix("http://www.w3.org/Style");',
+			message: messages.expected('url-prefix'),
+		},
+		{
+			code: '@document url-prefix( http://www.w3.org/Style );',
+			fixed: '@document url-prefix( "http://www.w3.org/Style" );',
+			message: messages.expected('url-prefix'),
+		},
+		{
+			code: '@document domain(mozilla.org);',
+			fixed: '@document domain("mozilla.org");',
+			message: messages.expected('domain'),
+		},
+		{
 			code: 'a { background: url(foo.css); }',
 			fixed: 'a { background: url("foo.css"); }',
+			message: messages.expected('url'),
+		},
+		{
+			code: 'a { background :   url(foo.css); }',
+			fixed: 'a { background :   url("foo.css"); }',
+			message: messages.expected('url'),
+			description: 'multiple spaces in between',
+		},
+		{
+			code: 'a { background-image: url(foo.css), url("bar.css"), url("baz.css"); }',
+			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url("baz.css"); }',
 			message: messages.expected('url'),
 		},
 	],
@@ -772,6 +819,16 @@ testRule({
 			description: 'multiple spaces after name',
 		},
 		{
+			code: "@document url-prefix('http://www.w3.org/Style');",
+			fixed: '@document url-prefix(http://www.w3.org/Style);',
+			message: messages.rejected('url-prefix'),
+		},
+		{
+			code: '@document domain("mozilla.org");',
+			fixed: '@document domain(mozilla.org);',
+			message: messages.rejected('domain'),
+		},
+		{
 			code: 'a { background: url("foo.css"); }',
 			fixed: 'a { background: url(foo.css); }',
 			message: messages.rejected('url'),
@@ -786,6 +843,11 @@ testRule({
 			fixed: 'a { background :   url(foo.css); }',
 			message: messages.rejected('url'),
 			description: 'multiple spaces in between',
+		},
+		{
+			code: "a { background-image: url('foo.css'), url(bar.css), url(baz.css); }",
+			fixed: 'a { background-image: url(foo.css), url(bar.css), url(baz.css); }',
+			message: messages.rejected('url'),
 		},
 	],
 });