--- a/lib/rules/function-url-quotes/__tests__/index.js
+++ b/lib/rules/function-url-quotes/__tests__/index.js
@@ -5,6 +5,7 @@ const { messages, ruleName } = require('..');
 testRule({
 	ruleName,
 	config: ['always'],
+	fix: true,
 
 	accept: [
 		{
@@ -180,6 +181,7 @@ testRule({
 	reject: [
 		{
 			code: '@import url(foo.css);',
+			fixed: '@import url("foo.css");',
 			message: messages.expected('url'),
 			line: 1,
 			column: 13,
@@ -188,6 +190,7 @@ testRule({
 		},
 		{
 			code: '@import url( foo.css );',
+			fixed: '@import url( "foo.css" );',
 			message: messages.expected('url'),
 			line: 1,
 			column: 14,
@@ -196,6 +199,7 @@ testRule({
 		},
 		{
 			code: '@document url-prefix(http://www.w3.org/Style);',
+			fixed: '@document url-prefix("http://www.w3.org/Style");',
 			message: messages.expected('url-prefix'),
 			line: 1,
 			column: 22,
@@ -204,6 +208,7 @@ testRule({
 		},
 		{
 			code: '@document url-prefix( http://www.w3.org/Style );',
+			fixed: '@document url-prefix( "http://www.w3.org/Style" );',
 			message: messages.expected('url-prefix'),
 			line: 1,
 			column: 23,
@@ -212,6 +217,7 @@ testRule({
 		},
 		{
 			code: "@font-face { font-family: 'foo'; src: url(foo.ttf); }",
+			fixed: '@font-face { font-family: \'foo\'; src: url("foo.ttf"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 43,
@@ -220,6 +226,7 @@ testRule({
 		},
 		{
 			code: "@font-face { font-family: 'foo'; src: url( foo.ttf ); }",
+			fixed: '@font-face { font-family: \'foo\'; src: url( "foo.ttf" ); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 44,
@@ -228,6 +235,7 @@ testRule({
 		},
 		{
 			code: 'a { cursor: url(foo.png); }',
+			fixed: 'a { cursor: url("foo.png"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 17,
@@ -236,6 +244,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url(foo.css), url("bar.css"), url("baz.css"); }',
+			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url("baz.css"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 27,
@@ -244,6 +253,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url( foo.css ), url("bar.css"), url("baz.css"); }',
+			fixed: 'a { background-image: url( "foo.css" ), url("bar.css"), url("baz.css"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 28,
@@ -252,6 +262,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url("foo.css"), url(bar.css), url("baz.css"); }',
+			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url("baz.css"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 43,
@@ -260,6 +271,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url("foo.css"), url( bar.css ), url("baz.css"); }',
+			fixed: 'a { background-image: url("foo.css"), url( "bar.css" ), url("baz.css"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 44,
@@ -268,6 +280,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url("foo.css"), url("bar.css"), url(baz.css); }',
+			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url("baz.css"); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 59,
@@ -276,6 +289,7 @@ testRule({
 		},
 		{
 			code: 'a { background-image: url("foo.css"), url("bar.css"), url( baz.css ); }',
+			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url( "baz.css" ); }',
 			message: messages.expected('url'),
 			line: 1,
 			column: 60,
@@ -288,6 +302,7 @@ testRule({
 testRule({
 	ruleName,
 	config: ['never'],
+	fix: true,
 
 	accept: [
 		{
@@ -397,6 +412,7 @@ testRule({
 	reject: [
 		{
 			code: '@import url("foo.css");',
+			fixed: '@import url(foo.css);',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 13,
@@ -405,6 +421,7 @@ testRule({
 		},
 		{
 			code: '@import uRl("foo.css");',
+			fixed: '@import uRl(foo.css);',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 13,
@@ -413,6 +430,7 @@ testRule({
 		},
 		{
 			code: '@import URL("foo.css");',
+			fixed: '@import URL(foo.css);',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 13,
@@ -421,6 +439,7 @@ testRule({
 		},
 		{
 			code: '@import url( "foo.css" );',
+			fixed: '@import url( foo.css );',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 14,
@@ -429,6 +448,7 @@ testRule({
 		},
 		{
 			code: "@import url('foo.css');",
+			fixed: '@import url(foo.css);',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 13,
@@ -437,6 +457,7 @@ testRule({
 		},
 		{
 			code: "@import url( 'foo.css' );",
+			fixed: '@import url( foo.css );',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 14,
@@ -445,6 +466,7 @@ testRule({
 		},
 		{
 			code: '@document url("http://www.w3.org/");',
+			fixed: '@document url(http://www.w3.org/);',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 15,
@@ -453,6 +475,7 @@ testRule({
 		},
 		{
 			code: '@document url( "http://www.w3.org/" );',
+			fixed: '@document url( http://www.w3.org/ );',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 16,
@@ -461,6 +484,7 @@ testRule({
 		},
 		{
 			code: "@document url-prefix('http://www.w3.org/Style');",
+			fixed: '@document url-prefix(http://www.w3.org/Style);',
 			message: messages.rejected('url-prefix'),
 			line: 1,
 			column: 22,
@@ -469,6 +493,7 @@ testRule({
 		},
 		{
 			code: "@document url-prefix( 'http://www.w3.org/Style' );",
+			fixed: '@document url-prefix( http://www.w3.org/Style );',
 			message: messages.rejected('url-prefix'),
 			line: 1,
 			column: 23,
@@ -477,6 +502,7 @@ testRule({
 		},
 		{
 			code: '@document domain("mozilla.org");',
+			fixed: '@document domain(mozilla.org);',
 			message: messages.rejected('domain'),
 			line: 1,
 			column: 18,
@@ -485,6 +511,7 @@ testRule({
 		},
 		{
 			code: '@document domain( "mozilla.org" );',
+			fixed: '@document domain( mozilla.org );',
 			message: messages.rejected('domain'),
 			line: 1,
 			column: 19,
@@ -493,6 +520,7 @@ testRule({
 		},
 		{
 			code: "@font-face { font-family: foo; src: url('foo.ttf'); }",
+			fixed: '@font-face { font-family: foo; src: url(foo.ttf); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 41,
@@ -501,6 +529,7 @@ testRule({
 		},
 		{
 			code: "@font-face { font-family: foo; src: url( 'foo.ttf' ); }",
+			fixed: '@font-face { font-family: foo; src: url( foo.ttf ); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 42,
@@ -509,6 +538,7 @@ testRule({
 		},
 		{
 			code: 'a { background: url("foo.css"); }',
+			fixed: 'a { background: url(foo.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 21,
@@ -517,6 +547,7 @@ testRule({
 		},
 		{
 			code: 'a { background: uRl("foo.css"); }',
+			fixed: 'a { background: uRl(foo.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 21,
@@ -525,6 +556,7 @@ testRule({
 		},
 		{
 			code: 'a { background: URL("foo.css"); }',
+			fixed: 'a { background: URL(foo.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 21,
@@ -533,6 +565,7 @@ testRule({
 		},
 		{
 			code: 'a { background: url( "foo.css" ); }',
+			fixed: 'a { background: url( foo.css ); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 22,
@@ -541,6 +574,7 @@ testRule({
 		},
 		{
 			code: 'a { background: url(  "foo.css"  ); }',
+			fixed: 'a { background: url(  foo.css  ); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 23,
@@ -549,6 +583,7 @@ testRule({
 		},
 		{
 			code: 'a { cursor: url("foo.png"); }',
+			fixed: 'a { cursor: url(foo.png); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 17,
@@ -557,6 +592,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url('foo.css'), url(bar.css), url(baz.css); }",
+			fixed: 'a { background-image: url(foo.css), url(bar.css), url(baz.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 27,
@@ -565,6 +601,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url( 'foo.css' ), url(bar.css), url(baz.css); }",
+			fixed: 'a { background-image: url( foo.css ), url(bar.css), url(baz.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 28,
@@ -573,6 +610,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url(foo.css), url('bar.css'), url(baz.css); }",
+			fixed: 'a { background-image: url(foo.css), url(bar.css), url(baz.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 41,
@@ -581,6 +619,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url(foo.css), url( 'bar.css' ), url(baz.css); }",
+			fixed: 'a { background-image: url(foo.css), url( bar.css ), url(baz.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 42,
@@ -589,6 +628,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url(foo.css), url(bar.css), url('baz.css'); }",
+			fixed: 'a { background-image: url(foo.css), url(bar.css), url(baz.css); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 55,
@@ -597,6 +637,7 @@ testRule({
 		},
 		{
 			code: "a { background-image: url(foo.css), url(bar.css), url( 'baz.css' ); }",
+			fixed: 'a { background-image: url(foo.css), url(bar.css), url( baz.css ); }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 56,
@@ -605,6 +646,7 @@ testRule({
 		},
 		{
 			code: 'a { background: url("/images/my_image@2x.png") }',
+			fixed: 'a { background: url(/images/my_image@2x.png) }',
 			message: messages.rejected('url'),
 			line: 1,
 			column: 21,
@@ -706,148 +748,6 @@ testRule({
 			column: 13,
 			endLine: 1,
 			endColumn: 22,
-		},
-	],
-});
-
-testRule({
-	ruleName,
-	config: ['always'],
-	fix: true,
-
-	reject: [
-		{
-			code: '@import url(foo.css);',
-			fixed: '@import url("foo.css");',
-			message: messages.expected('url'),
-		},
-		{
-			code: '@import url( foo.css );',
-			fixed: '@import url( "foo.css" );',
-			message: messages.expected('url'),
-		},
-		{
-			code: '@document url(http://www.w3.org/);',
-			fixed: '@document url("http://www.w3.org/");',
-			message: messages.expected('url'),
-		},
-		{
-			code: '@document url( http://www.w3.org/ );',
-			fixed: '@document url( "http://www.w3.org/" );',
-			message: messages.expected('url'),
-		},
-		{
-			code: '@document   url( http://www.w3.org/ );',
-			fixed: '@document   url( "http://www.w3.org/" );',
-			message: messages.expected('url'),
-			description: 'multiple spaces after name',
-		},
-		{
-			code: '@document url-prefix(http://www.w3.org/Style);',
-			fixed: '@document url-prefix("http://www.w3.org/Style");',
-			message: messages.expected('url-prefix'),
-		},
-		{
-			code: '@document url-prefix( http://www.w3.org/Style );',
-			fixed: '@document url-prefix( "http://www.w3.org/Style" );',
-			message: messages.expected('url-prefix'),
-		},
-		{
-			code: '@document domain(mozilla.org);',
-			fixed: '@document domain("mozilla.org");',
-			message: messages.expected('domain'),
-		},
-		{
-			code: 'a { background: url(foo.css); }',
-			fixed: 'a { background: url("foo.css"); }',
-			message: messages.expected('url'),
-		},
-		{
-			code: 'a { background :   url(foo.css); }',
-			fixed: 'a { background :   url("foo.css"); }',
-			message: messages.expected('url'),
-			description: 'multiple spaces in between',
-		},
-		{
-			code: 'a { background-image: url(foo.css), url("bar.css"), url("baz.css"); }',
-			fixed: 'a { background-image: url("foo.css"), url("bar.css"), url("baz.css"); }',
-			message: messages.expected('url'),
-		},
-	],
-});
-
-testRule({
-	ruleName,
-	config: ['never'],
-	fix: true,
-
-	reject: [
-		{
-			code: '@import url("foo.css");',
-			fixed: '@import url(foo.css);',
-			message: messages.rejected('url'),
-		},
-		{
-			code: '@import url( "foo.css" );',
-			fixed: '@import url( foo.css );',
-			message: messages.rejected('url'),
-		},
-		{
-			code: "@import url('foo.css');",
-			fixed: '@import url(foo.css);',
-			message: messages.rejected('url'),
-		},
-		{
-			code: "@import url( 'foo.css' );",
-			fixed: '@import url( foo.css );',
-			message: messages.rejected('url'),
-		},
-		{
-			code: '@document url("http://www.w3.org/");',
-			fixed: '@document url(http://www.w3.org/);',
-			message: messages.rejected('url'),
-		},
-		{
-			code: '@document url( "http://www.w3.org/" );',
-			fixed: '@document url( http://www.w3.org/ );',
-			message: messages.rejected('url'),
-		},
-		{
-			code: '@document   url( "http://www.w3.org/" );',
-			fixed: '@document   url( http://www.w3.org/ );',
-			message: messages.rejected('url'),
-			description: 'multiple spaces after name',
-		},
-		{
-			code: "@document url-prefix('http://www.w3.org/Style');",
-			fixed: '@document url-prefix(http://www.w3.org/Style);',
-			message: messages.rejected('url-prefix'),
-		},
-		{
-			code: '@document domain("mozilla.org");',
-			fixed: '@document domain(mozilla.org);',
-			message: messages.rejected('domain'),
-		},
-		{
-			code: 'a { background: url("foo.css"); }',
-			fixed: 'a { background: url(foo.css); }',
-			message: messages.rejected('url'),
-		},
-		{
-			code: "a { background: url('foo.css'); }",
-			fixed: 'a { background: url(foo.css); }',
-			message: messages.rejected('url'),
-		},
-		{
-			code: "a { background :   url('foo.css'); }",
-			fixed: 'a { background :   url(foo.css); }',
-			message: messages.rejected('url'),
-			description: 'multiple spaces in between',
-		},
-		{
-			code: "a { background-image: url('foo.css'), url(bar.css), url(baz.css); }",
-			fixed: 'a { background-image: url(foo.css), url(bar.css), url(baz.css); }',
-			message: messages.rejected('url'),
 		},
 	],
 });