--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -20,7 +20,7 @@ const meta = {
 };
 
 /** @type {import('stylelint').Rule} */
-const rule = (primary, secondaryOptions) => {
+const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
 		const validOptions = validateOptions(
 			result,
@@ -72,8 +72,53 @@ const rule = (primary, secondaryOptions) => {
 		}
 
 		/**
+		 * @param {import('postcss').Declaration | import('postcss').AtRule} node
 		 * @param {string} args
-		 * @param {import('postcss').Node} node
+		 * @param {number} index
+		 */
+
+		function addQuotes(node, args, index) {
+			const fixedName = `"${args}"`;
+			const openIndex = index - args.length - 1 - ('params' in node ? 0 : 4);
+			const closeIndex = openIndex + args.length;
+
+			if ('params' in node) {
+				node.params =
+					node.params.substring(0, openIndex) + fixedName + node.params.substring(closeIndex);
+
+				return;
+			}
+
+			node.value =
+				node.value.substring(0, openIndex) + fixedName + node.value.substring(closeIndex);
+		}
+
+		/**
+		 * @param {import('postcss').Declaration | import('postcss').AtRule} node
+		 * @param {string} args
+		 * @param {number} index
+		 */
+
+		function removeQuotes(node, args, index) {
+			// console.log(node)
+			// console.log(args)
+			// console.log(index)
+			const fixedName = args.substring(1, args.length - 1);
+			const openIndex = index - args.length + 1 - ('params' in node ? 0 : 4);
+			const closeIndex = openIndex + args.length;
+
+			if ('params' in node) {
+				node.params = node.params.slice(0, openIndex) + fixedName + node.params.slice(closeIndex);
+
+				return;
+			}
+
+			node.value = node.value.slice(0, openIndex) + fixedName + node.value.slice(closeIndex);
+		}
+
+		/**
+		 * @param {string} args
+		 * @param {import('postcss').Declaration | import('postcss').AtRule} node
 		 * @param {number} index
 		 * @param {string} functionName
 		 */
@@ -102,9 +147,21 @@ const rule = (primary, secondaryOptions) => {
 					return;
 				}
 
+				if (context.fix) {
+					addQuotes(node, trimmedArg, complaintIndex);
+
+					return;
+				}
+
 				complain(messages.expected(functionName), node, complaintIndex, complaintEndIndex);
 			} else {
 				if (!hasQuotes) {
+					return;
+				}
+
+				if (context.fix) {
+					removeQuotes(node, trimmedArg, complaintIndex);
+
 					return;
 				}