--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -19,6 +19,20 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-url-quotes',
 };
 
+/**
+ * @param {import('postcss').Declaration} decl
+ * @returns {number}
+ */
+function getDeclarationValueIndex(decl) {
+	let index = decl.prop.length;
+
+	if (decl.raws.between) {
+		index += decl.raws.between.length;
+	}
+
+	return index;
+}
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
@@ -79,7 +93,8 @@ const rule = (primary, secondaryOptions, context) => {
 
 		function addQuotes(node, args, index) {
 			const fixedName = `"${args}"`;
-			const openIndex = index - args.length - 1 - ('params' in node ? 0 : 4);
+			const openIndex =
+				'params' in node ? index - atRuleParamIndex(node) : index - getDeclarationValueIndex(node);
 			const closeIndex = openIndex + args.length;
 
 			if ('params' in node) {
@@ -100,11 +115,9 @@ const rule = (primary, secondaryOptions, context) => {
 		 */
 
 		function removeQuotes(node, args, index) {
-			// console.log(node)
-			// console.log(args)
-			// console.log(index)
-			const fixedName = args.substring(1, args.length - 1);
-			const openIndex = index - args.length + 1 - ('params' in node ? 0 : 4);
+			const fixedName = args.slice(1, args.length - 1);
+			const openIndex =
+				'params' in node ? index - atRuleParamIndex(node) : index - getDeclarationValueIndex(node);
 			const closeIndex = openIndex + args.length;
 
 			if ('params' in node) {