--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -63,7 +63,7 @@ const rule = (primary, secondaryOptions, context) => {
 		 * @param {import('postcss').Declaration} decl
 		 */
 		function checkDeclParams(decl) {
-			functionArgumentsSearch(decl.toString().toLowerCase(), 'url', (args, index) => {
+			functionArgumentsSearch(decl.toString(), /^url$/i, (args, index) => {
 				checkArgs(args, decl, index, 'url');
 			});
 		}
@@ -72,15 +72,13 @@ const rule = (primary, secondaryOptions, context) => {
 		 * @param {import('postcss').AtRule} atRule
 		 */
 		function checkAtRuleParams(atRule) {
-			const atRuleParamsLowerCase = atRule.params.toLowerCase();
-
-			functionArgumentsSearch(atRuleParamsLowerCase, 'url', (args, index) => {
+			functionArgumentsSearch(atRule.params, /^url$/i, (args, index) => {
 				checkArgs(args, atRule, index + atRuleParamIndex(atRule), 'url');
 			});
-			functionArgumentsSearch(atRuleParamsLowerCase, 'url-prefix', (args, index) => {
+			functionArgumentsSearch(atRule.params, /^url-prefix$/i, (args, index) => {
 				checkArgs(args, atRule, index + atRuleParamIndex(atRule), 'url-prefix');
 			});
-			functionArgumentsSearch(atRuleParamsLowerCase, 'domain', (args, index) => {
+			functionArgumentsSearch(atRule.params, /^domain$/i, (args, index) => {
 				checkArgs(args, atRule, index + atRuleParamIndex(atRule), 'domain');
 			});
 		}