--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const atRuleParamIndex = require('../../utils/atRuleParamIndex');
+const declarationValueIndex = require('../../utils/declarationValueIndex');
 const functionArgumentsSearch = require('../../utils/functionArgumentsSearch');
 const isStandardSyntaxUrl = require('../../utils/isStandardSyntaxUrl');
 const optionsMatches = require('../../utils/optionsMatches');
@@ -18,20 +19,6 @@ const messages = ruleMessages(ruleName, {
 const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-url-quotes',
 };
-
-/**
- * @param {import('postcss').Declaration} decl
- * @returns {number}
- */
-function getDeclarationValueIndex(decl) {
-	let index = decl.prop.length;
-
-	if (decl.raws.between) {
-		index += decl.raws.between.length;
-	}
-
-	return index;
-}
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
@@ -92,7 +79,7 @@ const rule = (primary, secondaryOptions, context) => {
 		function addQuotes(node, args, index) {
 			const fixedName = `"${args}"`;
 			const openIndex =
-				'params' in node ? index - atRuleParamIndex(node) : index - getDeclarationValueIndex(node);
+				'params' in node ? index - atRuleParamIndex(node) : index - declarationValueIndex(node);
 			const closeIndex = openIndex + args.length;
 
 			if ('params' in node) {
@@ -115,7 +102,7 @@ const rule = (primary, secondaryOptions, context) => {
 		function removeQuotes(node, args, index) {
 			const fixedName = args.slice(1, args.length - 1);
 			const openIndex =
-				'params' in node ? index - atRuleParamIndex(node) : index - getDeclarationValueIndex(node);
+				'params' in node ? index - atRuleParamIndex(node) : index - declarationValueIndex(node);
 			const closeIndex = openIndex + args.length;
 
 			if ('params' in node) {