@ybiquitous 2023-01-10
> First, this rule operates on both declarations and at-rules. It seems like in `checkArgs`, this type information is "erased". To resolve this, I have my utility functions `addQuotes` and `removeQuotes` take in a union type, and then use a crude `'params' in node` to differentiate the type.

Good point. I think it will be more readable to split into 2 functions with a single purpose. For example:

```js
if (node.type === 'atrule') {
	addQuotesForAtRule(node, trimmedArg, complaintIndex);
} else {
	addQuotesForDecl(node, trimmedArg, complaintIndex);
}
```

The same is true for `removeQuotes`.