@ybiquitous 2023-01-10
> To resolve this problem, I pass in a case-insensitive regex to `functionArgumentsSearch`. I don't think this is *technically* wrong, but it is likely a performance regression; I didn't want to change `functionArgumentsSearch` implementation instead. With that in mind, any suggestions on better ways to do this?

I think the solution makes sense and will not degrade performance much. But benchmarking is beneficial. Please consider sharing benchmark results via the `npm run benchmark-rule` script:

https://github.com/stylelint/stylelint/blob/b4fd2d50f588c5c4adb7f63a1091393b427f1a13/docs/developer-guide/rules.md?plain=1#L267