@Mouvedia 2023-01-12
@mattxwang does it use the `string-quotes` rule value—if not set to `null`—to choose between `"` and `'` for the fix?