@mattxwang 2023-01-14
> @mattxwang does it use the `string-quotes` rule value—if not set to `null`—to choose between `"` and `'` for the fix?
> 
> If not, that fix will require multiple runs. i.e. #3853

Ah - no, it doesn't. I implemented this similar to the behaviour in the other two quotes rules (`font-family-name-quotes`, `selector-attribute-quotes`) which both default to `"` as the quote. I believe I've added similar behaviour (i.e. in all three autofix cases, multiple runs are needed).

Is there past precedence / a blueprint for a rule inspecting the value of another rule? It may make sense to simply fix all of these at once in one PR (I am happy to contribute this if desired).