@Mouvedia 2023-01-14
> It may make sense to simply fix all of these at once in one PR (I am happy to contribute this if desired).

If you don't I probably will. Either way, you need an issue for it: TBD (lemme add it)